Properly remove views when animation completes
Pushing multiple views before the animation completed meant that only
the last view would get removed since we only stored the last item to
be removed. Instead, we store a stack of items to be removed and remove
the first on that stack when the animation completes. This also fixes
an issue where if the same view was pushed twice it will no longer
remove the active view when the animation completes.